Remove dependency from EmailCampaign on EmailAutomation
An email automation is a way to automatically generate email campaigns.
Email campaigns shouldn't need to know about the logic of the email
automations. The current setup makes it hard to reason about what is
happening in EmailCampaign because the behavior strongly depends on
whether automation is set or not.

With this commit EmailAutomation creates and triggers 2 separate
EmailCampaign objects, one for the admins and one for the end users
(unless it's admins_only).

This allows simplifying EmailCampaign. I will do that in a follow-up
commit to make reviewing the logic changes in this one easier to review.