Move rendering logic to EmailTemplate
EmailTemplate feels like the logical place for this because that's where the design, body, and subject are stored. The campaign should deal with the batching and sending and not care about how the email that is being sent is created. This will simplify the implementation of a configurable from_email/email sending server in a later commit.