Add email sender config class and set it as a field in EmailTemplate
An email sender config specifies the email server to use (including authentication settings) and the from_email to use for the emails. The email sender config is optional. If it is not set, the EmailBackend to use is constructed based on the values in the settings.py file (this is the current behavior).

I decided to make the EmailSenderConfig class a field in the EmailTemplate class because that is where currently all the data that is available in advance is specified for sending emails. One could argue that the email server information should be part of the EmailCampaign class, but then it would also need to be settable in the EmailAutomation class (because that creates EmailCampaigns) and that feels messy. Also, likely one would not want to send the same template via different email servers, so having it with the template doesn't hurt.